Linux/requirements.txt
Flask==2.0.3
matplotli
plotly
pandas
